backend/data_eng/requirements.txt
pandas
dask
distributed
mysql-connector
bs4
requests
dask
distributed
lxml
